+++ back/src/main/java/com/realive/repository/product/ProductImageRepository.java
@@ -1,0 +1,13 @@
+package com.realive.repository.product;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.realive.domain.product.ProductImage;
+
+public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
+    
+    List<ProductImage> findByProductId(Long productId);
+    
+}
